requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
chromadb==0.5.5
click==8.1.7
Flask==3.0.3
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
langchain==0.2.10
langchain-community==0.2.9
langchain-openai==0.1.17
lxml==5.2.2
pypdf==4.3.1
python-dotenv==1.0.1
MarkupSafe==2.1.5
Werkzeug==3.0.3
wikipedia==1.4.0
